backend/requirements.txt
flask
flask-cors
ultralytics
torch
numpy
pillow
jsonschema
opencv-python
